modified optics.psf_image output to be an ImageF, this is required (#34_optics_wrapper)

+++ galsim/optics.py
@@ -241,7 +241,7 @@ def psf_image(array_shape=(256, 256), defocus=0., astig1=0., astig2=0., coma1=0.
     """
     array = psf(array_shape=array_shape, defocus=defocus, astig1=astig1, astig2=astig2, coma1=coma1,
                 coma2=coma2, spher=spher, kmax=kmax, circular_pupil=circular_pupil, obs=obs)
-    return galsim.ImageF(array)
+    return galsim.ImageF(array.astype(np.float32))
 
 def otf(array_shape=(256, 256), defocus=0., astig1=0., astig2=0., coma1=0., coma2=0., spher=0.,
         kmax=np.pi, circular_pupil=True, obs=None):
